removed requirement to copy run_script
